executor_service/requirements.txt
fastapi
uvicorn
python-jose
fastapi
uvicorn
python-multipart
pydantic
sqlalchemy
psycopg2-binary
python-jose
python-dotenv
python-dotenv
praw